backend/requirements.txt
fastapi
uvicorn
websockets
openai-whisper
torch
soundfile
transformers
gtts
librosa
numpy 